// generated/src/app/api-models/claim-audit-mult-percent.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimAuditMultPercent {
  claimResultCode: string;
  payPercentage: number;
  procedureModifier: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}